chore(telemetry): trim 7 unused public exports, fold notice.ts into the entry, fix a stale comment
- index.ts no longer re-exports createTelemetry, DEFAULT_POSTHOG_KEY,
  CLOUD_PROP_KEYS, resolveConsent, saveConfig, configPath, maybeShowNotice —
  nothing in the repo, the console, or the examples imports them.
  EVENT_ALLOWLIST stays: fixtures/integration's telemetry-wire seam test
  consumes it.
- edge.ts sheds the same names (resolveConsent, no-op saveConfig) so the two
  entries' surfaces stay aligned.
- notice.ts (21 lines, single consumer) folded into index.ts as private code;
  its tests move to index.test.ts and now run through the public initTelemetry
  with a real temp home instead of injected callbacks.
- config.ts comment no longer cites a `vendo telemetry` command that has
  never existed.
